requirements.txt
discord.py
pymongo
disputils
jishaku
coloredlogs
yaml
colorthief
emoji